renamed nft dir
