--- nft/tezos_fa2_nft_tests/kill-sandbox.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-docker kill flextesa-sandbox
