--- nft/tezos_fa2_nft_tests/start-sandbox.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-docker run --rm --name flextesa-sandbox -e block_time=5 --detach -p 20000:20000 registry.gitlab.com/tezos/flextesa:image-tutobox-run carthagebox start
