+++ nft_assets/fa2_single_token.mligo
