+++ nft_assets/impl/fa2_single_asset.mligo
