+++ nft_assets/tezos_fa2_nft_tests/test_simple.py
